Add stack-path values as outputs
Use-cases are described in the new README section.

While we can dynamically walk over `stack path` and set and output for
every value we see in the Step, I can't find any way to avoid having to
maintain the static list at the Action level (and in the test case).