Automatically add --resolver nightly
It's not supported to put "resolver: nightly" into a stack.yaml; you
have to pick a specific "nightly-YYYY-MM-DD". This means you can't
simply run --stack-yaml stack-nightly.yaml and expect that to actually
run with an up-to-date nightly. You need to add --resolver nightly too.

This has lead to CI setups with a nightly job outside the main matrix,
since that job (and only that job) needs a special stack-arguments.

With this little bit of convenience, we can add our stack-nightly.yaml
into the main matrix and it Just Works.