Don't fall back to "any" cache
Separate the `stack.yaml` from the `package.yaml` one and only fall back
as far as the same `stack.yaml`. In multi-`stack-yaml` projects, which
are common, the old behavior could lead to a lot of cross-cache errors.
Therefore, taking it into account this way is a safer default.

NOTE: If `stack-arguments` should be taken into account, users are
expected to do that themselves (as the example workflow does).