Remove stack-yaml input
Users can either set `STACK_YAML` or `stack-arguments` to add
`--stack-yaml`, which each come with their own trade-offs depending on
the nature of the Job in which this action is used. Either way, a
dedicated input is not necessary or useful.

Before this commit, users who set `STACK_YAML` _and_ used
`inputs.stack-yaml` could get themselves in trouble if they weren't
careful.

I decided to keep the input for now, to avoid needing a major version
bump but:

- Documentation says it's deprecated
- It triggers a warning
- It results in naively pushing the value into `stack-arguments`, which
  is what the behavior was before, but might misbehave if the user is
  also setting `env.STACK_YAML` or already has `--stack-yaml` in
  `stack-arguments`, which was also the behavior before, I guess