Check for dirty-files post-build
If the `.cabal` or `.yaml.lock` files are checked into git, and the
build produces changes to them, that is an error. It means the user made
a change to `package.yaml` or `stack.yaml` and forgot to commit the
resulting change in the related file.

We'd like `stack-action` to catch this error when run on the PRs that
introduce it, since it's easy to miss and can result in bigger problems
later (e.g. when that lands on `main` and/or is released). To start,
we'll `warn`, but we may shift the default to `error` in the future.