Add -type f to generate-matrix default find(1) options
This ensures we only list regular files and not symbolic links. It may
be the case that a project symlinks `stack.yaml` to one of the specific
configurations. In such cases, only the actual files should generate
matrix elements. Generating one for `stack.yaml` would be redundant.

Projects that attach extra tasks (weeder, coverage) to only the
`stack.yaml` build should either not use a symlink, or adjust the `find`
input accordingly.