feat: add `color` input
`--no-terminal` is redundant, since Stack is not inferring the GitHub
Actions pseudo-terminal as a terminal anyway. I believe this is a
historical accident from when we originally wrote this action based on
some CircleCI scripts we were using before.

GitHub Actions can, however, render colors just fine. So we should
actually go the other way and force that instead.

This is particularly important for scanning error output.

Before:

![](https://files.pbrisbin.com/screenshots/screenshot.154736.png)

After:

![](https://files.pbrisbin.com/screenshots/screenshot.160892.png)

I realized during testing that defaulting `stack-arguments` is probably
not the correct way to do this stuff because it suffers from the classic
problem of customizations clobbering defaults when what you really want
is "in addition" behavior.

This didn't really matter before, since the `--no-terminal` thing was
not impactful, as mentioned, but that's not true of `color`. So I made
this one a dedicated option instead.